backend/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
pyarrow>=13.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# SAML (existing)
python3-saml>=1.15.0
